% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mmd_matrix}
\alias{mmd_matrix}
\title{Abundance-weighted MMD distance matrix}
\usage{
mmd_matrix(P, K)
}
\arguments{
\item{P}{Relative abundance matrix, samples in rows.}

\item{K}{Protein x protein kernel matrix.}
}
\description{
Abundance-weighted MMD distance matrix
}
