% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{rbf_kernel}
\alias{rbf_kernel}
\title{RBF Gram matrix from a ground-metric distance matrix}
\usage{
rbf_kernel(C, sigma = NULL)
}
\arguments{
\item{C}{Protein x protein distance matrix.}

\item{sigma}{Bandwidth; median off-diagonal distance if NULL.}
}
\description{
RBF Gram matrix from a ground-metric distance matrix
}
