% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{mmd_matrix}
\alias{mmd_matrix}
\title{Abundance-weighted MMD distance matrix}
\usage{
mmd_matrix(P, K)
}
\arguments{
\item{P}{Relative abundance matrix, samples in rows.}

\item{K}{Protein x protein kernel matrix.}
}
\value{
A samples x samples matrix of MMD distances.
}
\description{
Abundance-weighted MMD distance matrix
}
\examples{
P <- rbind(s1 = c(0.8, 0.2), s2 = c(0.1, 0.9))
K <- matrix(c(1, 0.2, 0.2, 1), 2)
mmd_matrix(P, K)
}
