% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{rbf_kernel}
\alias{rbf_kernel}
\title{RBF Gram matrix from a ground-metric distance matrix}
\usage{
rbf_kernel(C, sigma = NULL)
}
\arguments{
\item{C}{Protein x protein distance matrix.}

\item{sigma}{Bandwidth; median off-diagonal distance if NULL.}
}
\value{
A protein x protein RBF Gram matrix.
}
\description{
RBF Gram matrix from a ground-metric distance matrix
}
\examples{
C <- as.matrix(stats::dist(matrix(c(0, 0, 1, 0, 0, 1), ncol = 2,
                                  byrow = TRUE)))
rbf_kernel(C)
}
