% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{read_fasta}
\alias{read_fasta}
\title{Read protein sequences from a FASTA file}
\usage{
read_fasta(path)
}
\arguments{
\item{path}{Path to an amino-acid FASTA file.}
}
\value{
Named character vector, names taken from the first whitespace-delimited
token of each header.
}
\description{
Read protein sequences from a FASTA file
}
